# apps/web/src/lib/service-clients/service-properties/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "properties_service",
    "description": "",
    "termsOfService": "https://macro.com/terms",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/properties/definitions": {
      "get": {
        "tags": ["Properties"],
        "summary": "List property definitions with flexible filtering",
        "operationId": "list_properties",
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "description": "Filter by scope: 'user', 'team', 'system', or 'all'",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PropertyScope"
            }
          },
          {
            "name": "include_options",
            "in": "query",
            "description": "Whether to include property options in the response",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "for_entity_type",
            "in": "query",
            "description": "Filter properties applicable to a specific entity type",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/EntityType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Properties retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PropertyDefinitionResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": ["Properties"],
        "summary": "Create a new property definition",
        "operationId": "create_property_definition",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyDefinitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Property definition created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDefinition"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "403": {
            "description": "Team membership required for team scope"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/properties/definitions/{definition_id}": {
      "get": {
        "tags": ["Properties"],
        "summary": "Get a property definition by ID",
        "operationId": "get_property_definition",
        "parameters": [
          {
            "name": "definition_id",
            "in": "path",
            "description": "Property definition ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Property definition retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDefinition"
                }
              }
            }
          },
          "400": {
            "description": "Invalid property ID"
          },
          "404": {
            "description": "Property definition not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": ["Properties"],
        "summary": "Delete a property definition",
        "operationId": "delete_property_definition",
        "parameters": [
          {
            "name": "definition_id",
            "in": "path",
            "description": "Property definition ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Property definition deleted successfully"
          },
          "400": {
            "description": "Invalid property ID"
          },
          "403": {
            "description": "Cannot modify system properties"
          },
          "404": {
            "description": "Property definition not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/properties/definitions/{definition_id}/options": {
      "get": {
        "tags": ["Properties"],
        "summary": "Get all options for a property (for dropdowns)",
        "operationId": "get_property_options",
        "parameters": [
          {
            "name": "definition_id",
            "in": "path",
            "description": "Property definition ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Property options retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PropertyOption"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Property not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": ["Properties"],
        "summary": "Add a new option to a property dropdown",
        "operationId": "add_property_option",
        "parameters": [
          {
            "name": "definition_id",
            "in": "path",
            "description": "Property definition ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPropertyOptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Property option created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyOption"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "403": {
            "description": "Cannot modify system properties"
          },
          "404": {
            "description": "Property not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/properties/definitions/{definition_id}/options/{option_id}": {
      "delete": {
        "tags": ["Properties"],
        "summary": "Delete a property option",
        "operationId": "delete_property_option",
        "parameters": [
          {
            "name": "definition_id",
            "in": "path",
            "description": "Property definition ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "option_id",
            "in": "path",
            "description": "Property option ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Property option deleted successfully"
          },
          "400": {
            "description": "Invalid option ID"
          },
          "403": {
            "description": "Cannot modify system properties"
          },
          "404": {
            "description": "Property option not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "patch": {
        "tags": ["Properties"],
        "summary": "Update a property option in place (rename / recolor / reorder).",
        "description": "The option id is preserved, so the change is reflected on every entity that references it.",
        "operationId": "update_property_option",
        "parameters": [
          {
            "name": "definition_id",
            "in": "path",
            "description": "Property definition ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "option_id",
            "in": "path",
            "description": "Property option ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePropertyOptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Property option updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyOption"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "403": {
            "description": "Cannot modify system properties"
          },
          "404": {
            "description": "Property or option not found"
          },
          "409": {
            "description": "An option with that value already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/properties/entities/bulk": {
      "post": {
        "tags": ["Properties"],
        "summary": "Get properties for multiple entities in bulk (public endpoint with user auth)",
        "description": "Only returns properties for entities the user has view permission for.\nEntities without permission are silently omitted from the response.",
        "operationId": "get_bulk_entity_properties",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkEntityPropertiesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Bulk entity properties retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "$ref": "#/components/schemas/EntityPropertiesResponse"
                  },
                  "propertyNames": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "403": {
            "description": "Forbidden - User does not have permission"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/properties/entities/{entity_type}/{entity_id}": {
      "get": {
        "tags": ["Properties"],
        "summary": "Get all properties for an entity",
        "operationId": "get_entity_properties",
        "parameters": [
          {
            "name": "entity_type",
            "in": "path",
            "description": "Canonical entity type; tasks use DOCUMENT",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PropertyTargetEntityType"
            }
          },
          {
            "name": "entity_id",
            "in": "path",
            "description": "Entity ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include_metadata",
            "in": "query",
            "description": "Whether to include property metadata (default: false)",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entity properties retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityPropertiesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid entity type"
          },
          "403": {
            "description": "Forbidden - User does not have permission to view this entity"
          },
          "404": {
            "description": "Entity not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/properties/entities/{entity_type}/{entity_id}/options/bulk": {
      "post": {
        "tags": ["Properties"],
        "summary": "Apply a complete tag-picker selection across an entity's multi-select\nproperties in one request.",
        "description": "Each property change is expressed as an option delta and applied to the\ncurrent stored value under a per-row lock inside a single transaction, so the\nselection persists atomically and composes with concurrent edits instead of\nclobbering them. Returns the reconciled final option ids per property for\ncache reconciliation.",
        "operationId": "bulk_update_entity_property_options",
        "parameters": [
          {
            "name": "entity_type",
            "in": "path",
            "description": "Canonical entity type; tasks use DOCUMENT",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PropertyTargetEntityType"
            }
          },
          {
            "name": "entity_id",
            "in": "path",
            "description": "Entity ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkUpdateEntityPropertyOptionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Options updated; returns final ids per property",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkUpdateEntityPropertyOptionsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, a property is not multi-select, or an option does not belong to its property"
          },
          "403": {
            "description": "No edit access to the entity"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/properties/entities/{entity_type}/{entity_id}/{property_id}": {
      "put": {
        "tags": ["Properties"],
        "summary": "Set or update a property value for an entity, or attach a property without a value",
        "operationId": "set_entity_property",
        "parameters": [
          {
            "name": "entity_type",
            "in": "path",
            "description": "Canonical entity type; tasks use DOCUMENT",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PropertyTargetEntityType"
            }
          },
          {
            "name": "entity_id",
            "in": "path",
            "description": "Entity ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "property_id",
            "in": "path",
            "description": "Property ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetEntityPropertyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Entity property set successfully (with or without value)"
          },
          "400": {
            "description": "Invalid request or entity type"
          },
          "403": {
            "description": "No edit access to the entity"
          },
          "404": {
            "description": "Entity or property not found"
          },
          "409": {
            "description": "Task transition blocked by dependencies",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDependencyReadiness"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/properties/entities/{entity_type}/{entity_id}/{property_id}/options/{option_id}": {
      "post": {
        "tags": ["Properties"],
        "summary": "Add a single option to an entity's multi-select property value.",
        "description": "Atomic delta: the option is appended to the current stored value (deduped,\nattaching the property if needed), so concurrent edits to the same value\nmerge rather than overwrite each other. Prefer this over the full-value PUT\nwhen adding one option.",
        "operationId": "add_entity_property_option",
        "parameters": [
          {
            "name": "entity_type",
            "in": "path",
            "description": "Canonical entity type; tasks use DOCUMENT",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PropertyTargetEntityType"
            }
          },
          {
            "name": "entity_id",
            "in": "path",
            "description": "Entity ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "property_id",
            "in": "path",
            "description": "Property ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "option_id",
            "in": "path",
            "description": "Option ID to add",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Option added successfully"
          },
          "400": {
            "description": "Invalid request, property is not multi-select, or option does not belong to the property"
          },
          "403": {
            "description": "No edit access to the entity"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": ["Properties"],
        "summary": "Remove a single option from an entity's multi-select property value.",
        "description": "Atomic delta: the option is stripped from the current stored value (a no-op\nif absent), so concurrent edits to the same value merge rather than overwrite\neach other.",
        "operationId": "remove_entity_property_option",
        "parameters": [
          {
            "name": "entity_type",
            "in": "path",
            "description": "Canonical entity type; tasks use DOCUMENT",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PropertyTargetEntityType"
            }
          },
          {
            "name": "entity_id",
            "in": "path",
            "description": "Entity ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "property_id",
            "in": "path",
            "description": "Property ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "option_id",
            "in": "path",
            "description": "Option ID to remove",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Option removed successfully"
          },
          "403": {
            "description": "No edit access to the entity"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/properties/entity_properties/{entity_property_id}": {
      "delete": {
        "tags": ["Properties"],
        "summary": "Remove a specific entity property by its ID",
        "operationId": "delete_entity_property",
        "parameters": [
          {
            "name": "entity_property_id",
            "in": "path",
            "description": "Entity Property ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Entity property removed successfully"
          },
          "403": {
            "description": "Property is required and cannot be removed, or no edit access"
          },
          "404": {
            "description": "Entity property not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/properties/options/bulk": {
      "post": {
        "tags": ["Properties"],
        "summary": "Apply one shared option delta (add / remove option ids on a single\nmulti-select property) to many entities in one request — e.g. tag a set of\nemails with one label.",
        "description": "Best-effort per entity: one edit receipt is minted per entity, entities the\ncaller can't edit are reported as `skipped_no_permission` (mirroring the\nread path, which silently drops entities the caller can't view), and each\npermitted entity is updated in its own transaction so one entity failing\ndoes not roll back the others. Returns one result per requested entity in\nrequest order, with the reconciled final option ids for the successes.",
        "operationId": "bulk_update_entities_property_options",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkUpdateEntitiesPropertyOptionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Per-entity results: each entity applied, skipped, or failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkUpdateEntitiesPropertyOptionsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, the property is not multi-select, or an option does not belong to the property"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/properties/tags/merge": {
      "post": {
        "tags": ["Properties"],
        "summary": "Replace one of the caller's personal labels with an existing team label.",
        "description": "Every entity carrying the personal label is retagged with the team label\n(deduped if it already has both) and the personal label is deleted. The team\nlabel's name and color win. This is the confirmation step after the promote\nendpoint reports a name collision.",
        "operationId": "merge_tag",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeTagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Label merged into the team label",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyOptionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "403": {
            "description": "Team membership required"
          },
          "404": {
            "description": "Either label was not found in the expected tag set"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePropertyDefinitionRequest": {
        "type": "object",
        "description": "Request to create a new property definition.",
        "required": ["scope", "display_name", "data_type"],
        "properties": {
          "data_type": {
            "$ref": "#/components/schemas/PropertyDataType"
          },
          "display_name": {
            "type": "string"
          },
          "scope": {
            "$ref": "#/components/schemas/CreatePropertyScope"
          }
        }
      },
...
      "EntityType": {
        "type": "string",
        "description": "Type of entity that can be referenced by entity properties.",
        "enum": [
          "CALENDAR_EVENT",
          "CALL_RECORD",
          "CHANNEL",
          "CHAT",
          "COMPANY",
          "DOCUMENT",
          "PROJECT",
          "TASK",
          "THREAD",
          "USER"
        ]
      },
...
      "PropertyDefinitionResponse": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/PropertyDefinition"
          },
          {
            "$ref": "#/components/schemas/PropertyDefinitionWithOptions"
          }
        ],
        "description": "Response for property definition with optional property options"
      },
...
      "PropertyScope": {
        "type": "string",
        "description": "Scope filter for property queries",
        "enum": ["user", "team", "system", "all"]
      },
...
      "UpdatePropertyOptionRequest": {
        "type": "object",
        "description": "Request to update a property option in place. Omitted fields keep their current value.\nThe option id is preserved, so the change propagates to every entity referencing it.",
        "properties": {
          "color": {
            "type": ["string", "null"],
            "description": "New color as a hex string like `#RRGGBB`."
          },
          "display_order": {
            "type": ["integer", "null"],
            "format": "int32",
            "description": "New display order."
          },
          "value": {
            "type": ["string", "null"],
            "description": "New string value (label text). Only valid for SelectString and Tag properties."
          }
        }
      }
    }
  },
...
}